refactor: improve interface structuring
