--- src/main/java/mrquackduck/imageemojis/interfaces/ResourcePackGenerator.java
@@ -1,7 +1,0 @@
-package mrquackduck.imageemojis.interfaces;
-
-import mrquackduck.imageemojis.types.models.ResourcePack;
-
-public interface ResourcePackGenerator {
-    ResourcePack generate();
-}
