+++ src/main/java/mrquackduck/imageemojis/services/abstractions/IEmojiRepository.java
@@ -1,0 +1,9 @@
+package mrquackduck.imageemojis.services.abstractions;
+
+import mrquackduck.imageemojis.types.models.EmojiModel;
+
+import java.util.List;
+
+public interface IEmojiRepository {
+    List<EmojiModel> getEmojis();
+}
