+++ src/main/java/mrquackduck/imageemojis/services/abstractions/IHttpResourcePackServer.java
@@ -1,0 +1,6 @@
+package mrquackduck.imageemojis.services.abstractions;
+
+public interface IHttpResourcePackServer {
+    void start();
+    void stop();
+}
