+++ src/main/java/mrquackduck/imageemojis/services/abstractions/IResourcePackGenerator.java
@@ -1,0 +1,7 @@
+package mrquackduck.imageemojis.services.abstractions;
+
+import mrquackduck.imageemojis.types.models.ResourcePack;
+
+public interface IResourcePackGenerator {
+    ResourcePack generate();
+}
